Changed event sorting to a more general algorithm.
The previous sorter of new events could not sort more than two unordered
events, and that was only if they were at the first and second index.
The new one sorts any new given event into the correct place, assuming
that the already given list of events is already in order. Also, the
expected order has changed so that the oldest events are at the head of
the list, making any future event players more efficient with the data
structure.

The set_admin/1 function and admin field were removed from Funcard.Player.